aigent/apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Agentic Data Analytics Platform - Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=44.0.0",
    # Configuration
    "pydantic-settings>=2.7.0",
    # LangChain & AI
    "langchain>=0.3.14",
    "langgraph>=0.2.60",
    "langsmith>=0.2.10",
    "langchain-google-genai>=2.0.8",
    # WebSockets
    "websockets>=14.1",
    "pydantic>=2.12.5",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
